Make CorrelationLogContext an internal class to TestSerilogEvents.
