+++ serilog-utilities-concurrent-correlator/serilog-utilities-concurrent-correlator.csproj
@@ -50,7 +50,6 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="ConcurrentBagSink.cs" />
-    <Compile Include="CorrelationLogContext.cs" />
     <Compile Include="LoggerSinkConfigurationExtensions.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="TestSerilogLogEvents.cs" />
